The Home page was created, the welcome area was changed to a new Home page
