+++ src/pages/login/Login.jsx
@@ -4,17 +4,14 @@ import { useEffect } from "react";
 import "./Login.css";
 
 export default function Login() {
-  const { user, loginGoogleWithPopUp, logout, observeAuthState } =
+  const { user, loginGoogleWithPopUp, observeAuthState } =
     useAuthStore();
 
   const handleLogin = useCallback(() => {
     loginGoogleWithPopUp();
   }, [loginGoogleWithPopUp]);
 
-  const handleLogout = useCallback(() => {
-    logout();
-  }, [logout]);
-
+ 
   useEffect(() => {
     observeAuthState();
     console.log(user);
@@ -22,21 +19,12 @@ export default function Login() {
 
   return (
     <div className="container-login">
-      {user ? (
-        <>
-          <p className="welcome-text">Bienvenido, {user.displayName}</p>
-          <button className="button-logout" onClick={handleLogout}>
-            Cerrar Sesión
-          </button>
-        </>
-      ) : (
-        <>
-          <p className="welcome-text">Inicia sesión para continuar</p>
-          <button className="button-login" onClick={handleLogin}>
-            Iniciar Sesión
-          </button>
-        </>
-      )}
+      <>
+        <p className="welcome-text">Inicia sesión para continuar</p>
+        <button className="button-login" onClick={handleLogin}>
+          Iniciar Sesión
+        </button>
+      </>
     </div>
   );
 }
